move Chain::signature_targets to routing_table
